Change types used for storing atom length from int and size_t to uint64_t, so we can handle large files (large mdat atoms really), add new ignore method to BinaryStream

+++ source/MP4.AVCC.cpp
@@ -73,7 +73,7 @@ inline int createMask(int a, int b)
     return r;
 }
 
-void AVCC::processData( MP4::BinaryStream * stream, size_t length )
+void AVCC::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     std::streampos start = stream->tellg();
