+++ source/MP4.EDTS.cpp
@@ -18,7 +18,7 @@ std::string EDTS::getContent( void )
     return o.str();
 }
 
-void EDTS::processData( MP4::BinaryStream * stream, size_t length )
+void EDTS::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
