+++ source/MP4.ILOC.cpp
@@ -45,7 +45,7 @@ std::string ILOC::getContent( void )
     return o.str();
 }
 
-void ILOC::processData( MP4::BinaryStream * stream, size_t length )
+void ILOC::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
