+++ source/MP4.ILST.cpp
@@ -18,7 +18,7 @@ std::string ILST::getContent( void )
     return o.str();
 }
 
-void ILST::processData( MP4::BinaryStream * stream, size_t length )
+void ILST::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
