+++ source/MP4.MDAT.cpp
@@ -45,7 +45,7 @@ std::string MDAT::getContent( void )
     return o.str();
 }
 
-void MDAT::processData( MP4::BinaryStream * stream, size_t length )
+void MDAT::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
