+++ source/MP4.MFHD.cpp
@@ -45,7 +45,7 @@ std::string MFHD::getContent( void )
     return o.str();
 }
 
-void MFHD::processData( MP4::BinaryStream * stream, size_t length )
+void MFHD::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
