+++ source/MP4.MVEX.cpp
@@ -18,7 +18,7 @@ std::string MVEX::getContent( void )
     return o.str();
 }
 
-void MVEX::processData( MP4::BinaryStream * stream, size_t length )
+void MVEX::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
