+++ source/MP4.PADB.cpp
@@ -45,7 +45,7 @@ std::string PADB::getContent( void )
     return o.str();
 }
 
-void PADB::processData( MP4::BinaryStream * stream, size_t length )
+void PADB::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
