+++ source/MP4.PROF.cpp
@@ -45,7 +45,7 @@ std::string PROF::getContent( void )
     return o.str();
 }
 
-void PROF::processData( MP4::BinaryStream * stream, size_t length )
+void PROF::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
