+++ source/MP4.SCHI.cpp
@@ -45,7 +45,7 @@ std::string SCHI::getContent( void )
     return o.str();
 }
 
-void SCHI::processData( MP4::BinaryStream * stream, size_t length )
+void SCHI::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
