+++ source/MP4.STDP.cpp
@@ -45,7 +45,7 @@ std::string STDP::getContent( void )
     return o.str();
 }
 
-void STDP::processData( MP4::BinaryStream * stream, size_t length )
+void STDP::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
