+++ source/MP4.TFHD.cpp
@@ -45,7 +45,7 @@ std::string TFHD::getContent( void )
     return o.str();
 }
 
-void TFHD::processData( MP4::BinaryStream * stream, size_t length )
+void TFHD::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
