+++ source/MP4.TRAF.cpp
@@ -18,7 +18,7 @@ std::string TRAF::getContent( void )
     return o.str();
 }
 
-void TRAF::processData( MP4::BinaryStream * stream, size_t length )
+void TRAF::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     (void) stream;
     (void) length;
