+++ source/MP4.TREF.cpp
@@ -45,7 +45,7 @@ std::string TREF::getContent( void )
     return o.str();
 }
 
-void TREF::processData( MP4::BinaryStream * stream, size_t length )
+void TREF::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
