+++ source/MP4.TRUN.cpp
@@ -45,7 +45,7 @@ std::string TRUN::getContent( void )
     return o.str();
 }
 
-void TRUN::processData( MP4::BinaryStream * stream, size_t length )
+void TRUN::processData( MP4::BinaryStream * stream, uint64_t length )
 {
     stream->ignore( length );
 }
